+++ ios/Classes/MediaPickerPlusPlugin.swift
@@ -262,14 +262,14 @@ public class SwiftMediaPickerPlusPlugin: NSObject, FlutterPlugin, UIImagePickerC
     }
 
     private func hasGalleryPermission() -> Bool {
-        // TODO: Treat .limited access as granted.
-        return PHPhotoLibrary.authorizationStatus() == .authorized
+        let status = PHPhotoLibrary.authorizationStatus()
+        return status == .authorized || status == .limited
     }
 
     private func requestGalleryPermission(completion: @escaping (Bool) -> Void) {
         PHPhotoLibrary.requestAuthorization { status in
             DispatchQueue.main.async {
-                completion(status == .authorized)
+                completion(status == .authorized || status == .limited)
             }
         }
     }
@@ -1561,8 +1561,7 @@ public class SwiftMediaPickerPlusPlugin: NSObject, FlutterPlugin, UIImagePickerC
         }
         
         // Apply quality and save
-        // TODO: Avoid integer division; values < 100 become 0.
-        let quality = (options["imageQuality"] as? Int ?? 80) / 100
+        let quality = (options["imageQuality"] as? Int ?? 80) / 100.0
         let documentsDirectory = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first!
         let filename = "processed_\(Int(Date().timeIntervalSince1970)).jpg"
         let fileURL = documentsDirectory.appendingPathComponent(filename)
@@ -1610,8 +1609,7 @@ public class SwiftMediaPickerPlusPlugin: NSObject, FlutterPlugin, UIImagePickerC
         let watermarkedImage = addWatermark(to: image, text: watermarkText, fontSize: fontSize, position: position)
         
         // Save the watermarked image
-        // TODO: Avoid integer division; values < 100 become 0.
-        let quality = (options["imageQuality"] as? Int ?? 80) / 100
+        let quality = (options["imageQuality"] as? Int ?? 80) / 100.0
         let documentsDirectory = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first!
         let filename = "watermarked_image_\(Int(Date().timeIntervalSince1970)).jpg"
         let fileURL = documentsDirectory.appendingPathComponent(filename)